updated bounty to new bank and tests pass yay
