+++ pallets/bank/src/tests.rs
@@ -11,6 +11,7 @@ use sp_core::H256;
 use sp_runtime::{
     testing::Header,
     traits::IdentityLookup,
+    ModuleId,
     Perbill,
 };
 use util::{
@@ -35,6 +36,8 @@ impl_outer_event! {
         system<T>,
         pallet_balances<T>,
         org<T>,
+        vote<T>,
+        donate<T>,
         bank<T>,
     }
 }
@@ -91,15 +94,29 @@ impl org::Trait for Test {
     type Shares = u64;
     type ReservationLimit = ReservationLimit;
 }
+impl vote::Trait for Test {
+    type Event = TestEvent;
+    type VoteId = u64;
+    type Signal = u64;
+}
 parameter_types! {
-    pub const MinimumTransfer: u64 = 10;
+    pub const TransactionFee: u64 = 3;
+    pub const TreasuryModuleId: ModuleId = ModuleId(*b"py/trsry");
+}
+impl donate::Trait for Test {
+    type Event = TestEvent;
+    type Currency = Balances;
+    type TransactionFee = TransactionFee;
+    type Treasury = TreasuryModuleId;
+}
+parameter_types! {
+    pub const MaxTreasuryPerOrg: u32 = 50;
     pub const MinimumInitialDeposit: u64 = 20;
 }
 impl Trait for Test {
     type Event = TestEvent;
-    type BankId = u64;
     type Currency = Balances;
-    type MinimumTransfer = MinimumTransfer;
+    type MaxTreasuryPerOrg = MaxTreasuryPerOrg;
     type MinimumInitialDeposit = MinimumInitialDeposit;
 }
 pub type System = system::Module<Test>;
@@ -107,7 +124,7 @@ pub type Balances = pallet_balances::Module<Test>;
 pub type Org = org::Module<Test>;
 pub type Bank = Module<Test>;
 
-fn get_last_event() -> RawEvent<u64, u64, u64, u64> {
+fn get_last_event() -> RawEvent<u64, u64, u64> {
     System::events()
         .into_iter()
         .map(|r| r.event)
@@ -165,34 +182,21 @@ fn opening_bank_account_works() {
         let one = Origin::signed(1);
         let sixnine = Origin::signed(69);
         assert_noop!(
-            Bank::account_opens_account_for_org_with_deposit(
-                sixnine, 1, 10, None
-            ),
-            Error::<Test>::CannotOpenBankAccountForOrgIfNotOrgMember
+            Bank::open_org_bank_account(sixnine, 1, 10, None),
+            Error::<Test>::NotPermittedToOpenBankAccountForOrg
         );
         assert_noop!(
-            Bank::account_opens_account_for_org_with_deposit(
-                one.clone(),
-                1,
-                19,
-                None
-            ),
+            Bank::open_org_bank_account(one.clone(), 1, 19, None),
             Error::<Test>::CannotOpenBankAccountIfDepositIsBelowModuleMinimum
         );
         let total_bank_count = Bank::total_bank_count();
         assert_eq!(total_bank_count, 0u32);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
+        assert_ok!(Bank::open_org_bank_account(one.clone(), 1, 20, None));
         assert_eq!(
             get_last_event(),
-            RawEvent::AccountOpensOrgBankAccount(
+            RawEvent::BankAccountOpened(
                 1,
                 OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                1,
                 20,
                 1,
                 None
@@ -200,415 +204,5 @@ fn opening_bank_account_works() {
         );
         let total_bank_count = Bank::total_bank_count();
         assert_eq!(total_bank_count, 1u32);
-    });
-}
-
-#[test]
-fn account_2_org_transfer() {
-    new_test_ext().execute_with(|| {
-        let one = Origin::signed(1);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        let two = Origin::signed(2);
-        assert_noop!(
-            Bank::account_to_org_transfer(
-                two.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 0]),
-                19,
-            ),
-            Error::<Test>::TransferFailsIfDestBankDNE
-        );
-        assert_noop!(
-            Bank::account_to_org_transfer(
-                two.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                9,
-            ),
-            Error::<Test>::TransferMustExceedModuleMinimum
-        );
-        assert_noop!(
-            Bank::account_to_org_transfer(
-                two.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                100,
-            ),
-            DispatchError::Module {
-                index: 0,
-                error: 3,
-                message: Some("InsufficientBalance")
-            }
-        );
-        assert_ok!(Bank::account_to_org_transfer(
-            two.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            10,
-        ));
-        assert_eq!(
-            get_last_event(),
-            RawEvent::AccountToOrgTransfer(
-                2,
-                2,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                10
-            ),
-        );
-    });
-}
-
-#[test]
-fn org_2_acc_transfer_from_transfer() {
-    new_test_ext().execute_with(|| {
-        let one = Origin::signed(1);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        // send 10 from 2 to org 1
-        let two = Origin::signed(2);
-        assert_ok!(Bank::account_to_org_transfer(
-            two.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            10,
-        ));
-        // anything greater than 10 is not allowed
-        assert_noop!(
-            Bank::org_to_account_transfer_from_transfer(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                2,
-                2,
-                11,
-            ),
-            Error::<Test>::CannotTransferFromOrgToAccountIfInWrongStateOrNotEnoughFunds
-        );
-        // send 5 back from org 1 to 2
-        assert_ok!(Bank::org_to_account_transfer_from_transfer(
-            one.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            2,
-            2,
-            5,
-        ));
-        let transfer_id = TransferId::new(OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]), 2);
-        // sharp state transition so spends are no longer allowed
-        assert_ok!(
-            Bank::stop_spends_start_withdrawals(transfer_id)
-        );
-        // and it is enforced
-        assert_noop!(
-            Bank::org_to_account_transfer_from_transfer(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                2,
-                2,
-                5,
-            ),
-            Error::<Test>::CannotTransferFromOrgToAccountIfInWrongStateOrNotEnoughFunds
-        );
-    });
-}
-
-#[test]
-fn org_2_org_transfer_from_transfer() {
-    new_test_ext().execute_with(|| {
-        let one = Origin::signed(1);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        // cannot transfer to self
-        assert_noop!(
-            Bank::org_to_org_transfer_from_transfer(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                1,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                10
-            ),
-            Error::<Test>::ThisModuleDoesNotPermitTransfersToSelf
-        );
-        // register second bank account
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        assert_eq!(
-            get_last_event(),
-            RawEvent::AccountOpensOrgBankAccount(
-                1,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-                1,
-                20,
-                1,
-                None
-            ),
-        );
-        // cannot transfer more than amount in transfer
-        assert_noop!(
-            Bank::org_to_org_transfer_from_transfer(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                1,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-                21
-            ),
-            Error::<Test>::CannotTransferFromOrgToOrgIfInWrongStateOrNotEnoughFunds
-        );
-        // can transfer 10 because this is less than or equal to 20
-        assert_ok!(Bank::org_to_org_transfer_from_transfer(
-            one.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            1,
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-            10
-        ));
-        assert_eq!(
-            get_last_event(),
-            RawEvent::OrgToOrgTransferFromTransfer(TransferId { id: OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]), sub_id: 2 }, OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]), OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]), 10),
-        );
-    });
-}
-
-#[test]
-fn reserve_spend_for_acc() {
-    new_test_ext().execute_with(|| {
-        let one = Origin::signed(1);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        assert_noop!(
-            Bank::reserve_spend_for_account(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                1,
-                2,
-                21
-            ),
-            Error::<Test>::ReserveOrgSpendExceedsFreeTransferCapital
-        );
-        assert_noop!(
-            Bank::reserve_spend_for_account(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                2,
-                2,
-                5
-            ),
-            Error::<Test>::CannotReserveOrgSpendIfTransferDNE
-        );
-        assert_ok!(Bank::reserve_spend_for_account(
-            one.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            1,
-            2,
-            5
-        ),);
-        assert_eq!(
-            get_last_event(),
-            RawEvent::ReserveAccountSpendFromTransfer(
-                TransferId {
-                    id: OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                    sub_id: 1
-                },
-                TransferId {
-                    id: OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                    sub_id: 2
-                },
-                2,
-                5
-            ),
-        );
-    });
-}
-
-#[test]
-fn reserve_spend_for_org() {
-    new_test_ext().execute_with(|| {
-        let one = Origin::signed(1);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        assert_noop!(
-            Bank::reserve_spend_for_org(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                1,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-                5
-            ),
-            Error::<Test>::TransferFailsIfDestBankDNE
-        );
-        // register second bank account
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        assert_noop!(
-            Bank::reserve_spend_for_org(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 3]),
-                1,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-                5
-            ),
-            Error::<Test>::CannotReserveOrgSpendIfBankStoreDNE
-        );
-        assert_noop!(
-            Bank::reserve_spend_for_org(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                1,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-                21
-            ),
-            Error::<Test>::ReserveOrgSpendExceedsFreeTransferCapital
-        );
-        assert_noop!(
-            Bank::reserve_spend_for_org(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                2,
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-                5
-            ),
-            Error::<Test>::CannotReserveOrgSpendIfTransferDNE
-        );
-        assert_ok!(Bank::reserve_spend_for_org(
-            one.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            1,
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-            5
-        ));
-        assert_eq!(
-            get_last_event(),
-            RawEvent::ReserveOrgSpendFromTransfer(
-                TransferId {
-                    id: OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                    sub_id: 1
-                },
-                TransferId {
-                    id: OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                    sub_id: 2
-                },
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 2]),
-                5
-            ),
-        );
-    });
-}
-
-#[test]
-fn transfer_reserved_spend() {
-    new_test_ext().execute_with(|| {
-        let one = Origin::signed(1);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        assert_ok!(Bank::reserve_spend_for_account(
-            one.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            1,
-            2,
-            5
-        ),);
-        assert_noop!(
-            Bank::transfer_existing_reserved_spend(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                3,
-                6,
-            ),
-            Error::<Test>::TransferReservedFailsIfSpendReservationDNE
-        );
-        // odd how the first reservation is 2 but the second id is 3 so it's ok
-        assert_noop!(
-            Bank::transfer_existing_reserved_spend(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                2,
-                6,
-            ),
-            Error::<Test>::TransferReservedFailsIfSpendReservationAmtIsLessThanRequest
-        );
-        assert_ok!(
-            Bank::transfer_existing_reserved_spend(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                2,
-                5,
-            )
-        );
-        assert_eq!(
-            get_last_event(),
-            RawEvent::OrgToAccountReservedSpendExecuted(TransferId { id: OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]), sub_id: 2 }, 2, 5),
-        );
-        // second reserved spend
-        assert_ok!(Bank::reserve_spend_for_account(
-            one.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            1,
-            2,
-            5
-        ),);
-        assert_ok!(
-            Bank::transfer_existing_reserved_spend(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                3,
-                5,
-            )
-        );
-        assert_eq!(
-            get_last_event(),
-            RawEvent::OrgToAccountReservedSpendExecuted(TransferId { id: OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]), sub_id: 3 }, 2, 5),
-        );
-    });
-}
-
-#[test]
-fn withdraw_from_org_to_acc() {
-    new_test_ext().execute_with(|| {
-        let one = Origin::signed(1);
-        assert_ok!(Bank::account_opens_account_for_org_with_deposit(
-            one.clone(),
-            1,
-            20,
-            None
-        ));
-        assert_ok!(Bank::stop_transfers_to_trigger_withdrawals(
-            one.clone(),
-            OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-            1
-        ));
-        assert_noop!(
-            Bank::withdraw_from_org_to_account(
-                one.clone(),
-                OnChainTreasuryID([0, 0, 0, 0, 0, 0, 0, 1]),
-                1
-            ),
-            Error::<Test>::TransferNotInValidStateToMakeRequestedWithdrawal
-        );
     });
 }
