--- pallets/bank2/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "sunshine-bank2"
-version = "0.0.5"
-authors = ["Amar Singh <asinghchrony@protonmail.com>"]
-edition = "2018"
-
-license = "GPL-3.0"
-repository = "https://github.com/sunshine-protocol/sunshine-bounty"
-description = "bank accounts for orgs (n treasury limit per org)"
-keywords = ["sunshine", "substrate"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false, features = ["derive"] }
-pallet-balances = { version = "2.0.0-rc4", default-features = false }
-sp-std = { version = "2.0.0-rc4", default-features = false }
-sp-runtime = { version = "2.0.0-rc4", default-features = false }
-frame-support = { version = "2.0.0-rc4", default-features = false }
-frame-system = { version = "2.0.0-rc4", default-features = false }
-util = { package = "sunshine-util", path = "../../pallets-util", default-features=false}
-org = { package = "sunshine-org", path = "../org", default-features=false}
-vote = { package = "sunshine-vote", path = "../vote", default-features=false}
-donate = { package = "sunshine-donate", path = "../donate", default-features=false}
-clear_on_drop = { version = "0.2.4", features = ["no_cc"] } # https://github.com/paritytech/substrate/issues/4179
-# temp until ./bank.rs is merged into pallets-util
-derive-new = { version = "0.5.8", default-features = false }
-sp-core = { version = "2.0.0-rc4", default-features = false }
-
-[dev-dependencies]
-pallet-balances = { version = "2.0.0-rc4", default-features = false }
-sp-io = { version = "2.0.0-rc4", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-    "codec/std",
-    "sp-std/std",
-    "sp-core/std",
-    "sp-runtime/std",
-    "frame-support/std",
-    "frame-system/std",
-    "pallet-balances/std",
-    "org/std",
-]
