+++ pallets/bounty/Cargo.toml
@@ -20,6 +20,7 @@ pallet-balances = { version = "2.0.0-rc4", default-features = false }
 util = { package = "sunshine-util", path = "../../pallets-util", default-features=false}
 org = { package = "sunshine-org", path = "../org", default-features=false}
 vote = { package = "sunshine-vote", path = "../vote", default-features=false}
+donate = { package = "sunshine-donate", path = "../donate", default-features=false}
 bank = { package = "sunshine-bank", path = "../bank", default-features=false}
 clear_on_drop = { version = "0.2.4", features = ["no_cc"] }	# https://github.com/paritytech/substrate/issues/4179
