+++ pallets/bounty/src/tests.rs
@@ -11,6 +11,7 @@ use sp_core::H256;
 use sp_runtime::{
     testing::Header,
     traits::IdentityLookup,
+    ModuleId,
     Perbill,
 };
 
@@ -32,6 +33,7 @@ impl_outer_event! {
         pallet_balances<T>,
         org<T>,
         vote<T>,
+        donate<T>,
         bank<T>,
         bounty<T>,
     }
@@ -95,15 +97,23 @@ impl vote::Trait for Test {
     type Signal = u64;
 }
 parameter_types! {
-    // minimum deposit to register an on-chain bank
-    pub const MinimumInitialDeposit: u64 = 5;
-    pub const MinimumTransfer: u64 = 5;
+    pub const TransactionFee: u64 = 3;
+    pub const TreasuryModuleId: ModuleId = ModuleId(*b"py/trsry");
+}
+impl donate::Trait for Test {
+    type Event = TestEvent;
+    type Currency = Balances;
+    type TransactionFee = TransactionFee;
+    type Treasury = TreasuryModuleId;
+}
+parameter_types! {
+    pub const MaxTreasuryPerOrg: u32 = 50;
+    pub const MinimumInitialDeposit: u64 = 20;
 }
 impl bank::Trait for Test {
     type Event = TestEvent;
-    type BankId = u64;
     type Currency = Balances;
-    type MinimumTransfer = MinimumTransfer;
+    type MaxTreasuryPerOrg = MaxTreasuryPerOrg;
     type MinimumInitialDeposit = MinimumInitialDeposit;
 }
 parameter_types! {
@@ -120,7 +130,7 @@ pub type Balances = pallet_balances::Module<Test>;
 pub type Org = org::Module<Test>;
 pub type Bounty = Module<Test>;
 
-fn get_last_event() -> RawEvent<u64, u64, u64, u64, u64> {
+fn get_last_event() -> RawEvent<u64, u64, u64, u64> {
     System::events()
         .into_iter()
         .map(|r| r.event)
@@ -545,9 +555,7 @@ fn account_sudo_approves_milestone_works() {
                 1,
                 1,
                 1,
-                MilestoneStatus::ApprovedAndTransferExecuted(
-                    BankOrAccount::Account(2)
-                )
+                MilestoneStatus::ApprovedAndTransferExecuted,
             )
         );
     });
